Extract shared GeneratorEvent base to remove duplication (#137)
SonarCloud flagged the new GeneratorPreBuyEvent as duplicated code: it
repeated the targetPlayer/islandUUID/generator/generatorID fields and
accessors that GeneratorBuyEvent, GeneratorActivationEvent and
GeneratorUnlockEvent already duplicated among themselves.

Introduce an abstract GeneratorEvent base holding those common fields and
their getters/setters, and have all four events extend it. The change is
backwards compatible: the public accessors are simply inherited, so
external listeners are unaffected. Each concrete event keeps its own
HandlerList, as Bukkit requires for correct per-type routing.
