+++ src/main/java/world/bentobox/magiccobblestonegenerator/events/GeneratorActivationEvent.java
@@ -7,12 +7,9 @@
 package world.bentobox.magiccobblestonegenerator.events;
 
 
-import java.util.UUID;
-
 import org.bukkit.event.Cancellable;
 import org.bukkit.event.HandlerList;
 
-import world.bentobox.bentobox.api.events.BentoBoxEvent;
 import world.bentobox.bentobox.api.user.User;
 import world.bentobox.magiccobblestonegenerator.database.objects.GeneratorTierObject;
 
@@ -20,7 +17,7 @@ import world.bentobox.magiccobblestonegenerator.database.objects.GeneratorTierOb
 /**
  * This event is fired when user is trying to activate or deactivate generator. It is cancellable.
  */
-public class GeneratorActivationEvent extends BentoBoxEvent implements Cancellable
+public class GeneratorActivationEvent extends GeneratorEvent implements Cancellable
 {
     /**
      * Instantiates a new Generator activation event.
@@ -32,101 +29,8 @@ public class GeneratorActivationEvent extends BentoBoxEvent implements Cancellab
      */
     public GeneratorActivationEvent(GeneratorTierObject generator, User user, String island, boolean activate)
     {
-        this.generator = generator.getFriendlyName();
-        this.generatorID = generator.getUniqueId();
-
-        this.targetPlayer = user.getUniqueId();
-        this.islandUUID = island;
-
+        super(generator, user, island);
         this.activate = activate;
-    }
-
-
-    /**
-     * Gets target player.
-     *
-     * @return the target player
-     */
-    public UUID getTargetPlayer()
-    {
-        return targetPlayer;
-    }
-
-
-    /**
-     * Sets target player.
-     *
-     * @param targetPlayer the target player
-     */
-    public void setTargetPlayer(UUID targetPlayer)
-    {
-        this.targetPlayer = targetPlayer;
-    }
-
-
-    /**
-     * Gets island uuid.
-     *
-     * @return the island uuid
-     */
-    public String getIslandUUID()
-    {
-        return islandUUID;
-    }
-
-
-    /**
-     * Sets island uuid.
-     *
-     * @param islandUUID the island uuid
-     */
-    public void setIslandUUID(String islandUUID)
-    {
-        this.islandUUID = islandUUID;
-    }
-
-
-    /**
-     * Gets generator.
-     *
-     * @return the generator
-     */
-    public String getGenerator()
-    {
-        return generator;
-    }
-
-
-    /**
-     * Sets generator.
-     *
-     * @param generator the generator
-     */
-    public void setGenerator(String generator)
-    {
-        this.generator = generator;
-    }
-
-
-    /**
-     * Gets generator id.
-     *
-     * @return the generator id
-     */
-    public String getGeneratorID()
-    {
-        return generatorID;
-    }
-
-
-    /**
-     * Sets generator id.
-     *
-     * @param generatorID the generator id
-     */
-    public void setGeneratorID(String generatorID)
-    {
-        this.generatorID = generatorID;
     }
 
 
@@ -209,26 +113,6 @@ public class GeneratorActivationEvent extends BentoBoxEvent implements Cancellab
 // ---------------------------------------------------------------------
 // Section: Variables
 // ---------------------------------------------------------------------
-
-    /**
-     * Player who activates generator.
-     */
-    private UUID targetPlayer;
-
-    /**
-     * Island Id.
-     */
-    private String islandUUID;
-
-    /**
-     * Friendly name for generator.
-     */
-    private String generator;
-
-    /**
-     * Generator ID.
-     */
-    private String generatorID;
 
     /**
      * Boolean that indicates if generator will be activated or deactivated.
