+++ src/main/java/world/bentobox/magiccobblestonegenerator/events/GeneratorPreBuyEvent.java
@@ -7,12 +7,9 @@
 package world.bentobox.magiccobblestonegenerator.events;
 
 
-import java.util.UUID;
-
 import org.bukkit.event.Cancellable;
 import org.bukkit.event.HandlerList;
 
-import world.bentobox.bentobox.api.events.BentoBoxEvent;
 import world.bentobox.bentobox.api.user.User;
 import world.bentobox.magiccobblestonegenerator.database.objects.GeneratorTierObject;
 
@@ -22,7 +19,7 @@ import world.bentobox.magiccobblestonegenerator.database.objects.GeneratorTierOb
  * requirements to the generator purchasing process. Cancelling this event stops the purchase before any money is
  * withdrawn and before {@link GeneratorBuyEvent} is fired.
  */
-public class GeneratorPreBuyEvent extends BentoBoxEvent implements Cancellable
+public class GeneratorPreBuyEvent extends GeneratorEvent implements Cancellable
 {
     /**
      * Instantiates a new Generator pre buy event.
@@ -33,99 +30,7 @@ public class GeneratorPreBuyEvent extends BentoBoxEvent implements Cancellable
      */
     public GeneratorPreBuyEvent(GeneratorTierObject generator, User user, String island)
     {
-        this.generator = generator.getFriendlyName();
-        this.generatorID = generator.getUniqueId();
-
-        this.targetPlayer = user.getUniqueId();
-        this.islandUUID = island;
-    }
-
-
-    /**
-     * Gets target player.
-     *
-     * @return the target player
-     */
-    public UUID getTargetPlayer()
-    {
-        return targetPlayer;
-    }
-
-
-    /**
-     * Sets target player.
-     *
-     * @param targetPlayer the target player
-     */
-    public void setTargetPlayer(UUID targetPlayer)
-    {
-        this.targetPlayer = targetPlayer;
-    }
-
-
-    /**
-     * Gets island uuid.
-     *
-     * @return the island uuid
-     */
-    public String getIslandUUID()
-    {
-        return islandUUID;
-    }
-
-
-    /**
-     * Sets island uuid.
-     *
-     * @param islandUUID the island uuid
-     */
-    public void setIslandUUID(String islandUUID)
-    {
-        this.islandUUID = islandUUID;
-    }
-
-
-    /**
-     * Gets generator.
-     *
-     * @return the generator
-     */
-    public String getGenerator()
-    {
-        return generator;
-    }
-
-
-    /**
-     * Sets generator.
-     *
-     * @param generator the generator
-     */
-    public void setGenerator(String generator)
-    {
-        this.generator = generator;
-    }
-
-
-    /**
-     * Gets generator id.
-     *
-     * @return the generator id
-     */
-    public String getGeneratorID()
-    {
-        return generatorID;
-    }
-
-
-    /**
-     * Sets generator id.
-     *
-     * @param generatorID the generator id
-     */
-    public void setGeneratorID(String generatorID)
-    {
-        this.generatorID = generatorID;
+        super(generator, user, island);
     }
 
 
@@ -186,26 +91,6 @@ public class GeneratorPreBuyEvent extends BentoBoxEvent implements Cancellable
 // ---------------------------------------------------------------------
 // Section: Variables
 // ---------------------------------------------------------------------
-
-    /**
-     * Player who buys generator Id.
-     */
-    private UUID targetPlayer;
-
-    /**
-     * Island Id.
-     */
-    private String islandUUID;
-
-    /**
-     * Friendly name for generator.
-     */
-    private String generator;
-
-    /**
-     * Generator ID.
-     */
-    private String generatorID;
 
     /**
      * Boolean that indicates if event is cancelled.
